Setup projection data for mkDat

Setup projection data for mkDat

setup.projDat(mesh, projection.dat, expert.formula, gating.formula)

Arguments

mesh

spatial constrained Delaunay triangulation derived from the fmesher R package

projection.dat

Spatial Points class of projection coordinates or Spatial Points Dataframe containing projection coordinates and projection covariates

expert.formula

Formula defining expert model. This formula corresponds to the covariates included in the response densities. Defaults to intercept only (~1) when no covariates are used.

gating.formula

Formula defining gating model. This formula corresponds to the covariates included in the mixing proportions (logistic regression). Defaults to intercept only (~1) when no covariates are used. When a random effects term is included in the gating network, this formula will be updated so that the intercept term is removed.

Value

list of projection data